requirements.txt
opentelemetry-instrumentation-logging
opentelemetry-semantic-conventions